+++ src/main/java/br/com/zalf/prolog/webservice/services/RelatorioService.java
@@ -11,7 +11,7 @@ public class RelatorioService {
 	private RelatorioDaoImpl dao = new RelatorioDaoImpl();
 
 	public ConsolidadoHolder getRelatorioByPeriodo(LocalDate dataInicial, LocalDate dataFinal, String equipe,
-			int codUnidade, Long cpf, String token){
+			Long codUnidade, Long cpf, String token){
 		try{
 			return dao.getRelatorioByPeriodo(dataInicial, dataFinal, equipe, codUnidade, cpf, token);
 		}catch (SQLException e) {
